Converted tests to use CommonJS format.
