+++ brjs-sdk/sdk/libs/javascript/br-formatting/tests/test-unit/js-test-driver/jsTestDriver.conf
@@ -6,4 +6,4 @@ load:
   - bundles/js/js.bundle
 
 test:
-  - tests/br/**.js
+  - tests/**.js
