--- brjs-sdk/sdk/libs/javascript/br-formatting/tests/test-unit/js-test-driver/tests/br/presenter/formatter/UpperCaseFormatterTest.js
@@ -1,23 +1,0 @@
-UpperCaseFormatterTest = TestCase("UpperCaseFormatterTest");
-
-require('br/presenter/formatter/UpperCaseFormatter');
-
-UpperCaseFormatterTest.prototype.setUp = function() {
-	this.oFormatter = new br.presenter.formatter.UpperCaseFormatter();
-};
-
-UpperCaseFormatterTest.prototype.test_lower = function() {
-	assertEquals(" THIS IS A SENTENCE ", this.oFormatter.format(" this is a sentence "));
-};
-
-UpperCaseFormatterTest.prototype.test_upper = function() {
-	assertEquals(" THIS IS A SENTENCE ", this.oFormatter.format(" THIS IS A SENTENCE "));
-};
-
-UpperCaseFormatterTest.prototype.test_mixed = function() {
-	assertEquals(" THIS IS A SENTENCE ", this.oFormatter.format(" This is a Sentence "));
-};
-
-UpperCaseFormatterTest.prototype.test_toString = function() {
-	assertEquals("string", typeof this.oFormatter.toString());
-};
